wlgen/workload: fix execution into cgroups
The CGroups mount point is defined by the devlib's cgroup module which
exposes also a new method to properly format the required shutils command.
This patch makes use of this new method thus ensuring to always issue
a correct and complete shutils command to run a workload under the specifed
group.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>