wlgen/rtapp: fix loading of custom JSON
When 'kind=custom' the 'params' parameter must be used to specify the
full path of a JSON file to be used for the rt-app confiuration.
This feature is currently broken because we anticipate in the
rtapp::conf() method some checks required only by the "profile" kind.

This patch fixes this issue and improve some sanity checks on the values
of the specified parameters.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>