Add smoke tests for wlgen RT-App classes
These are just trivial smoke tests, essentially copied from the WlGenUsage
tutorial notebook, the bare minimum to avoid issues like
https://github.com/ARM-software/lisa/issues/253.

As a hack to allow running tests without a target device we use target=None and
an empty calibration dict.

To make the tests really trivial I have just used hard-coded values for the
expected JSON phase output, the rationale being that this avoids bugs in the
test code.

If we do more hacking on the wlgen classes then we should be able to add more
interesting tests here.