test_wlgen: Add smoke tests for wlgen classes
This adds basic tests for some wlgen classes. The functionality tested here is
basically what is contained in the current example notebooks, i.e. it just tests
that the API is not completely broken.

Some of the infrastructure added here can probably be used to self-test other
parts of LISA using a local target.

To make the tests really trivial I have just used hard-coded values for the
expected JSON phase output, the rationale being that this avoids bugs in the
test code.

If we do more hacking on the wlgen classes then we should be able to add more
interesting tests here.