base: Don't error on event field values containing '='
If a trace field value contains a '=', we currently get a ValueError('Too many
values to unpack'). Instead, let's only split on the first '='.

In practice if a field value contains a '=' it's probably because a kernel
developer typo'd a custom event like:

    /* Note missing space between "%d" and "bar" */
    trace_printk("my_broken_event: foo=%dbar=%d", foo, bar)

So I did consider raising an explicit 'malformed event field' error. But this
approach is more flexible in case someone really wanted to trace fields
containing strings with '=' in them.