utils/energy.py: call superclass init with res_dir
For the ACME energy probe the init funtion call of the superclass
did not specify the results directory so the sample file would go
by default to /tmp/, where it might be overwritten if two tests are
ran at the same time, where the channels used have the same channel
descriptors.

For safety, pass the res_dir to the init call and have the output
file stored there before being moved to the specified output folder.

Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>