utils/executor: fix task naming for backward compatibility
The executor has got the capability to generate multiple tasks based on
a single spec since this patch:

   43a28826 libs/utils/executor: Allow "profile" rt-app wloads to specify a "tasks" param

However, that implementation adds an additional numerical ID to all task
names, event when just one has to be generated.
This is overkilling, if just one task has to be generated, and it also
breaks existing tests which expect the exact name specified in the
wload specs.

Let fix this by appending a numerical ID only for tasks which wload specs
have specified a multiplicity. Moreover, let's better format the numerical
ID by prefixing it with an "_".

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>