utils/test: alias test configuration and workloads
Test configuration and workloads can be specified either using inline dicts
or JSON files. Moreover, they can be part of the same main dictionary or
separate ones. Finally, these information are tracked by different modules.

This patch adds two convenience alias which allows a test to get an easy
access to the set of configurations and workloads used to run the experiments.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>