uibench: test_name suffix broke Transition tests
test_name had a forced suffix of 'Activity' with the assumption that all
test names would have this appended at the end in the app. This is not
true for two activities; ActivityTransition and
ActivityTransitionDetails.
As such, the suffix has been removed and the user now must specify the
full name of the test name. In addition, the default list of tests have
been updated to reflect this.

Also updated ipynb and py files that use uibench.py