uibench: Added parameter to allow action(s)
By default when running uibench, no user action is performed during the
tests. This adds an optional parameter called 'action' and performs the
actions as input events. This parameter can take the form of a list and
allows you to perform multiple actions in a set order too.

E.g. When performing ActivityTransition test, action=['tap', 'tap'] will
perform two taps in the centre of the screen which will maximise a
thumbnail then minimise the resulting fullscreen image.