uibench: Set the duration when using systrace
uibench by default collects with systrace but the duration_s parameter
is not used correctly. This is because of a limitation that we cannot
CTRL+C stop systrace captures so the workload class expects duration to
be appended to the collect parameter. This fixes uibench workload to
correctly parse the collect parameter to include the given duration_s
parameter.