LisaTest: Improve running tests with multiple iterations
Currently when you run a LisaTest with experiments_conf['iterations']!=0, the
test assertions are made for every experiment and the test is failed as soon as
any experiment produces a failure. Only the first failure encountered is
displayed.

With this patch all the results for each wload/conf pair are aggregated.

The problem is only solved across iterations: if multiple wloads or confs fail,
only the first wload/conf pair that was found to have failed will be
displayed.