tests/eas/capacity_capping: Make assertions relative to wload start time
This test currently assumes when making assertions on traces that the workload
started at time 0. This is probably not the case as there is a gap between
starting ftrace and running the workload.

Use SchedMultiAssert.getStartTime to find the true start time of the workload
and then use this as a basis for migration windows when making residency
assertions.