tests/eas/capacity-capping: Don't let target latency influence phase times
This test assumes that the artificial constraints transitioned at specific
times, and uses  `time.sleep` to space out these transitions. However in between
these `time.sleep` calls it writes values on the target, which can take
significant time.

Measure the latency of this write, then subtract that latency from the time to
sleep.