uibench: default actions added as a dictionary
Default actions for every activity has been added as a dictionary.
The test_name activity is used to look up the default actions to be
performed and performs them during the test.

uibench: delay default and position changed

The delay between actions now occurs before the action, not after
The default delay has changed from 0 to 1