gfxinfo: Poll framestats in a thread
gfxinfo can be executed with an additional 'framestats' parameter that
gives detailed timestamp information for the last 120 frames.
Added a new parameter to expose this functionality.
As this only captures the last 120 frames, to get all frame data across
the workload we need to poll at least every 2 seconds (assuming 60Hz
vsync) and append the data to the same file. New start and stop
functions for doing this have been added to the workload class, similar
to how tracing is currently handled.