TestEnv: Don't error if target's energy model is invalid
It may happen that a target's EAS energy model is "invalid", e.g. because its
active state power values are not monotonically increasing with frequency (this
is the case with Google Pixel). In this case the EnergyModel construct will
throw a ValueError.

Don't fail to start LISA in that case, just print a warning.