energy/AEP: Use pandas to simplify reading energy data
Instead of using devlib's MeasurementsCsv directly, just read the CSV file into
a DataFrame and use that. The format of the CSV is specified by the devlib API
so this should be generic.

This loses the extra information that is currently stored in the EnergyCounter
object, however that object is not currently exposed by the API and is
undocumented. So let's remove it. When we need that information again, we can
get it from pandas much more easily.

Remove the use of `time`, as it assumes that the energy meter can be started and
stopped instantaneously, this is not trustworthy.