utils/android: let systrace use the default device
A common use-case is to have a single device connected to your
host machine and use a notebook to run some experiments.
If such a notebook does not specify a device ID we currently raise
an exception because we try to access a non existing configuration
property.

Let the systrace_start API silently use the default device if
one if not specified by the target configuration.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>