utils/android: add support to show last collected trace
A common usage pattern is to visualise a trace, being either in
ftrace or systrace format, which has been collected during the
last execution of a workload.

This adds a commodity method to the Android::Workload class which
allows to visualise the most recently collected trace using the
appropriate external visualization tool, i.e.
- kernelshark for ftrace
- the default browser for systrace

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>