executor: Allow providing Workload object directly
The Executor class currently requires you to use a dictionary grammar to
configure the workloads you want to run. This is not necessary when you are
using Python interactively and have direct access to the wlgen API. With this
commit, Executor gains the ability for the "wloads" field in the
experiments_conf to map names directly to workload objects.

A very simple test case is added to exercise at least one code path using this
new feature.