energy/AEP: Don't require channel_map
The ARM Energy Probe can only measure one channel at a time, and the
`labels` param just sets the key that's used to identify this sole
channel in the output. So if we don't get a channel_map arg, just
leave `labels` None and use whatever devlib calls it by default.