utils/trace: remove tasks maps building at trace parsing time
The current Trace ctor allows to specify a limited set of tasks of interest
with the idea that further analysis methods can work on these tasks only.

However, this functionality has never been really used and we have now
a more efficient and complete dictionary of all tasks available in a trace:
 - trace._tasks_by_name
 - trace._tasks_by_pid
with convenient get methods:
 - trace.getTaskByName()
 - trace.getTaskByPid()

This patch remove the building of the task names dictionary which is left
as an on-demand API to be completely removed by a following patch.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>