utils/trace: remove tasks attribute
The Trace object is tracking a set of tasks which can be used as a default
for other trace analysis modules. However, this API is somehow confusing
since most of the times we still want to specify a different set of tasks
after the trace has been parsed.

This patch removes the usage of the tasks attributes at Trace parsing time
while ensuring that each of the current API clients define which tasks
they want to focus on.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>