libs/utils/env.py: Add gem5 support
Gem5 is now a board option for TestEnv configuration.
board_settings is used to define the simulation

Results directory was being initialized after reading the
configurations and setting up the targets. This initialization
has been moved up in the code so that is it made before reading
the configurations, the result path is thus now available for
the initialization of certain boards (e.g. gem5)