doc: Use git to get version string
This uses git but falls back to "UNKNOWN_VERSION" if anything goes
wrong (e.g. we aren't in a git repo). The rationale for this is that the person
building the docs is proabably a LISA developer - they are very likely to be
using git and should be checking the log from the doc build for errors.