gem5: add platform information from platform file
If a platform file in the format gem5<description>.json
(where description is the platform description: e.g. juno)
exists in lib/utils/platforms/ add the platform information
from that file to the platform (Gem5SimulationPlatform)
instance.

When the platform file does not exist, devlib should also
be able to obtain some platform information (core_clusters)
from the filesystem once gem5 is able to properly report
the part information for the CPUs.

Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>