libs/utils: Add remote target script class
This class is used to accumulate commands that would
otherwise be sent to a genuine devlib target. This is helpful
to plan a series of commands to be later executed on a target
and avoid a lot of connectivity overhead. The class can then
be used to create a script containing all of those commands,
push it on the target and run it.

Also quoting Brendan Jackman on this: "a component that needs
to run a load of commands unconditionally in one go can use
this to avoid network/ADB round trips".