utils/trace: fix duplicated indexes for wakeup/idle events
In some traces it could happen that a CPU wakeup is immediately followed
by a CPU idle event (or the other way around) which have the exact same
timestamp. Such a condition generates an exception when we try to re-index
the active signal, e.g. in HeavyLoadTest::test_tasks_spread.

This fixes this issue by ensuring that the cpu_active time series does not
have duplicated indexes.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>