Update some code for Pandas v0.20.1
This Pandas release changes the returned format of groupby.describe
[1]. Update sme code snippets to handle this change. Essentially, we
no longer need to call `unstack` on the result.

In the tuturial notebook, assume the user has just installed LISA and
so has the latest pandas, and just remove the unstack call (currently
in commented code) completey to keep things simple.

In the Jankbench notebook, add a comment so that existing users who
have not updated their pandas have a line to uncomment if their
notebook stops working after they pull in this patch (which they
probably won't do).

In library code, try to work with both output formats.

Ideally we should have a proper solution for depency management but I
don't think any contributors have the required knowledge of Python
distribution "best practices" at the moment, and I don't think it's
worth spending the time to learn right now.

[1] http://pandas.pydata.org/pandas-docs/version/0.20/whatsnew.html#groupby-describe-formatting