android/workload: allow collection of kernel's symbols address
Kernel symbols are useful to support debugging of tasks traces and
ftrace allows to inject a backtrack from code via a simple call
to trace_dump_stack(). When the trace is dumped, the backtrace addresses
can be translated to function names via the mapping available in:
   /proc/kallsyms

However, for security reasons, the content of this file reports the
actual addresses only if access has not been "not restricted", which is
the default configuration.

This patch disable such restriction by pocking the provided procfs
configuration file (/proc/sys/kernel/kptr_restrict) every time we start
a tracing using either ftrace or systrace.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>