rta: Save and restore governor tunables in calibration
Governor tunables are reset to default when switching governors.

This is particularly relevant when using the Executor, which calls
te.calibration() after _setup_cpufreq. If the calibration has not
been done yet, that will result in setting the performance governor
then restoring the previous governor. However if _setup_cpufreq set
up any tunables they are not currently restored. This commit fixes
that.