energy_model: Fix read-multiple-files helper
This helper executes a "cat" command to read a large number of files
in one go. In order to avoid this command becomind too long for ADB
to cope with, it was written to use glob patterns to cat the files,
instead of the explicit paths (which it knows).

However a user has reported an issue with this approach, which I
suspect is because the order of the paths returned by `find` differs
from the order of the shell's expansion.

Re-use the find command to ensure that the files are read in the same
order. Also add a check for the files and file-contents having the
same lenght, and a warning in the docstring about the stability of
this function.