energy_model: Remove assertion on frequency coherency
This assert says that we don't get different frequencies within the
same clock domain. In when estimating energy from traces, we will
find that while the cpu_frequency events for an update are still
being issued, there will be a brief period while this is not true. So
just drop the assertion.