tests/eas/capacity_capping: add support for freezing userspace tasks
A high level of tasks running during this test can cause significant
disturbance as this test ensures that during a simulated thermal capping
event less than 1% of the time is spent running tasks on the cpus which
have been capped.

In the case that the uncapped cpus are very busy, it is still reasonable
to put the test tasks on the capped cpus which counts as a failure but
is probably the correct schedule.

Now that we have moved freezing into TestEnv, we can make direct
use of it.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>