trace: Add abstraction for CPU load-tracking trace events
There is a new upstream-friendly version 'sched_load_se' of the
load-tracking trace events added by EAS. Add an abstraction so that
LISA can work with both this new version and the older
'sched_load_avg_cpu' events.

Trace._sanitize_SchedLoadAvgCpu is removed and the logic is moved
into the new function _dfg_cpu_load_events in tasks_analysis.

The users of sched_load_avg_cpu in cpus_analysis are updated to use
the new _dfg method.

Also add basic unit tests to exercise the code changed in this
commit.