tests/generic: use 16ms period for all workloads
How "big" a periodic task is depends not only from its duty-cycle but
also from the period.
A 20% periodic task, can be considered "small" from the scheduler time
perspective if it's period is around the scheduler latency. However it's
completely different if its period if of 500ms. In this case indeed the
task runs for 100ms at each activation, thus it should better be
considered quite big.

In our synthetics tests, let's fix the period of all the generated
workloads such that how big a task is can be properly defined by just
setting its duty-cycle. We pick 16ms because it's the most interesting
time constant for a standard Android rendering pipeline.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>