env: Add cpufreq to default modules
We need the cpufreq module to configure rt-app. We don't currently
have an easy way to detect in advance that we need to do this
calibration, so for now just add cpufreq as a default module. This is
already done for all the known boards, this commit just makes it a
default for unkonwn boards too.