utils/trace: add cgroup_attach_task event sanitization
This patch aims to build a cgroup_attach_task data frame that is a combination
of fake devlib events (if available), and real cgroup_task_attach events. We
pass along cgroup information to the trace object to help with this (controller
id to name mapping and cgroup names). Some out of tree kernel patches already
have 'cgroup' field, so we don't add cgroup information to those and just add
controller information.

Signed-off-by: Joel Fernandes <joelaf@google.com>