energy: refactor base class for CONTINUOUS devlib instrument
This splits the monolitic report() method of the base class for
CONTINUOUS devlib instrument in several sub-methods that can be
individually re-defined in sub-classes. This can be needed to account
for caveats of new instruments, e.g. the gem5 power instrument, for
which *almost* all of report()'s code is the same.