tests/eas: add a base class for CPU hotplug stress-tests
The CPU hotplug procedure changes many data structures used by the scheduler
and interferes with the task placement, hence resulting in all sorts of bugs.
This commit adds a base class enabling testing under CPU hotplug stress as
an attemp to trigger those bugs. CPUs are plugged in and off in a random (but
repeatable) way while a workload is running. For now, the test is considered
passing if the workload simply completes its work and if the system doesn't
halt.