base: Improve errors for failures to parse
Some events from the kernel cannot be parsed by old trace-cmd
binaries. This results in data_str starting with "[FAILED TO PARSE]",
which we ourselves fail to parse because we expect to find a
key=value pair.

We currently fail with an akward KeyError when trying to use None as
a key for data_dict when that happens. This commit changes it so we
instead raise one of two explicit errors:

- If we see "FAILED TO PARSE", tell the user to update trace-cmd

- If something else has gone wrong, tell the user that TRAPpy is
  broken. I haven't seen this happen, but let's add the error so that
  if/when it does, we can easily tell what happened.