libs/utils/android/screen: Add screenrecord utility
screenrecord is an Android utility that can be used to record
a video of the device screen.

Timestamping can be overlayed on top of the video, which can make
trace analysis a bit easier: if the video shows stuttering, the
timestamp can be used to find out what was going on in the scheduler
at that time.